finish search Bar, ready to be integrated
