+++ src/components/SearchBar.js
@@ -1,70 +1,73 @@
 import Autosuggest from 'react-autosuggest';
-import React from 'react'
+import React from 'react';
 import './css/SearchBar.css'
 import axios from 'axios';
- 
+
 const renderSuggestion = suggestion => (
-  <div className="SearchBar--suggestion">
+  <div>
     {suggestion}
   </div>
 );
 
 class SearchBar extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      value: '',
-      suggestions: [],
-      aoc : [],
-    };
-    this.getAoc = this.getAoc.bind(this)
-  }
-  async getAoc(){
-    const response = await axios.get('https://plateforme.api-agro.fr/api/records/1.0/search/?dataset=delimitation-parcellaire-des-aoc-viticoles&facet=appellatio&facet=denominati&facet=crinao');
+    constructor(props) {
+        super(props);
+        this.state = {
+        value: '',
+        suggestions: [],
+        aoc : []
+        };
+    }
+    async componentDidMount(){
+        const response = await axios.get('https://plateforme.api-agro.fr/api/records/1.0/search/?dataset=delimitation-parcellaire-des-aoc-viticoles&facet=appellatio&facet=denominati&facet=crinao')
+        this.setState({
+         aoc : response.data.records
+    })
+    }
+
+    goodValue(){
+      return this.state.aoc.map((region) => { 
+        return region.fields.appellatio
+      }
+      )}
+
+    onChange = (event, { newValue }) => {
       this.setState({
-        aoc : response.data.records
+        value: newValue
       });
-  };
-    
-  getSuggestions(value) {
-    const inputValue = value.trim().toLowerCase();
-    const inputLength = inputValue.length;
-    return inputLength === 0 ? [] : this.state.aoc.filter(region =>
-      region.toLowerCase().slice(0, inputLength) === inputValue
-    );
-  };
-  getSuggestionValue = suggestion => suggestion;
-
-
-  onChange = (event, { newValue }) => {
-    this.getAoc()
-    this.setState({
-      value: newValue
-    });
-  };
-
-
-  onSuggestionsFetchRequested = ({ value }) => {
-    this.setState({
-      suggestions: this.getSuggestions(value)
-    });
-  };
-
-  onSuggestionsClearRequested = () => {
-    this.setState({
-      suggestions: []
-    });
-  };
-
-  render() {
-    const { value, suggestions } = this.state;
-  
-    const inputProps = {
-      placeholder: 'Entrez votre AOC',
-      value,
-      onChange: this.onChange
+      this.goodValue()
     };
 
+  
+
+    onSuggestionsFetchRequested = ({ value }) => {
+      this.setState({
+        suggestions: this.getSuggestions(value)
+      });
+    };
+
+    getSuggestions = value => {
+      const inputValue = value.trim().toLowerCase();
+      const inputLength = inputValue.length;
+      return inputLength === 0 ? [] : this.goodValue(this.state.aoc).filter(lang =>
+        lang.toLowerCase().slice(0, inputLength) === inputValue
+      );
+    };
+
+    onSuggestionsClearRequested = () => {
+      this.setState({
+        suggestions: []
+      });
+    };
+    getSuggestionValue = suggestion => suggestion;
+
+    render() {
+      const { value, suggestions } = this.state;
+      const inputProps = {
+        placeholder: 'Tape ton AOC',
+        value,
+        onChange: this.onChange
+      };
 
     return (
       <Autosuggest
@@ -78,7 +81,5 @@ class SearchBar extends React.Component {
     );
   }
 }
-
-
 
 export default SearchBar
